requirements.txt
flask
flask-cors
python-dotenv
pydantic
google-generativeai
google-adk[a2a]
